+++ internal/impl/parquet/processor_decode.go
@@ -95,11 +95,7 @@ func (s *parquetDecodeProcessor) Process(ctx context.Context, msg *service.Messa
 		return nil, err
 	}
 
-	pRdrConf, err := parquet.NewReaderConfig()
-	if err != nil {
-		return nil, err
-	}
-	pRdr := parquet.NewReader(inFile, pRdrConf)
+	pRdr := parquet.NewGenericReader[any](inFile)
 
 	rowBuf := make([]parquet.Row, 10)
 	var resBatch service.MessageBatch